slight changes from deleteProduct

+++ BackEnd/controller/product.controller.js
@@ -48,11 +48,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req,res) => {
     const {id} = req.params
     
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success: false, message: "Invalid product id"}); // 400 status code if client error
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted successfully"});
     } catch (error) {
         console.error("Error in deleting product:", error.message)
-        res.status(404).json({success: false, message: "Product not found"}); // 404 status code if not found
+        res.status(500).json({success: false, message: "Product not found"}); // 404 status code if not found
     }
 };
